+++ pkg/util/constants/profiles.go
@@ -28,10 +28,11 @@ const ProfilesDeployment = ProfileGroup + "/deployment"
 const ProfilesIntegrationPrefix = "integration." + ProfileGroup
 
 const (
-	ProfilesIntegrationAuthn = "authn"
-	ProfilesIntegrationAuthz = "authz"
-	ProfilesIntegrationSched = "sched"
-	ProfilesIntegrationEnvoy = "envoy"
+	ProfilesIntegrationAuthn   = "authn"
+	ProfilesIntegrationAuthz   = "authz"
+	ProfilesIntegrationSched   = "sched"
+	ProfilesIntegrationEnvoy   = "envoy"
+	ProfilesIntegrationStorage = "storage"
 )
 
 const (